+++ lib/core_image/filter.dart
@@ -4,6 +4,7 @@ class _CIFilter extends CIFilter {
   final String name;
   final int index;
   final _CIFilterGroup group;
+  Map<String, Map<String, String>> _attributes;
 
   _CIFilter(this.name, this.index, this.group);
 
@@ -15,7 +16,8 @@ class _CIFilter extends CIFilter {
     for (var key in inputKeys) {
       attributes[key] = await group._inputKeyDetails(this, key);
     }
-    return attributes;
+    _attributes = attributes;
+    return _attributes;
   }
 
   @override
@@ -25,11 +27,36 @@ class _CIFilter extends CIFilter {
   Future<void> export(File output) => group.export(output);
 
   @override
-  Future<void> setAssetSource(String name)  => group.setAssetSource(name);
+  Future<void> setAssetSource(String name) => group.setAssetSource(name);
 
   @override
   Future<void> setFileSource(File path) => group.setFileSource(path);
 
   @override
   Future<void> setSource(Uint8List data) => group.setSource(data);
+
+  @override
+  Future<void> setScalarValue(String key, double value) async {
+    var attributes = _attributes;
+    if (attributes == null) {
+      attributes = await this.attributes;
+    }
+    final properties = attributes[key];
+    if (properties  == null) {
+      return Future.error('$key is not acceptable for $name');
+    }
+    if (properties['CIAttributeType'] != 'CIAttributeTypeScalar') {
+      return Future.error('$key is not CIAttributeTypeScalar');
+    }
+    if (properties['CIAttributeClass'] != 'NSNumber') {
+      return Future.error('$key is not number format');
+    }
+    final min = double.parse(properties['CIAttributeSliderMin']);
+    final max = double.parse(properties['CIAttributeSliderMax']);
+
+    if (value < min || value > max) {
+      return Future.error('$value must be in range [$min, $max]');
+    }
+    return group._setScalarValue(this, key, value);
+  }
 }