+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+# native_filters
 
-* TODO: Describe initial release.
+## 0.0.1+alpha - 1 Jun 2020
+
+* Added CIFilter wrapper 
+* Added platform views for filters demo
+* Implemented video and image processing.
+* Added grouping CIFilters